Search item from appbar and fix routes for the login
Now you can search for a item using the appbar, by two ways; typing and using a list view, or using the cam to read a barcode.
Fixing the route for the navigation to the login page, once logged you can't back to the login page, and when you log out, all the pages in the stack behind are destroyed.