UI and UX change on the table of stock for the item.
Now tapping on the number of the "other stock" column in the table of stock, you can see a new table with the stock of the specific item, in the other locals.
Add a image for the result "empty" of the images of the item.
Changes on the user model to add seller id and local.
On the services of item stock, we send a new parameter to specify what type of stock are we looking for, "local" or "others".
On the services of user, we prepare the lecture of the response for the re-structured user.