The inherited widget is implemented for user.
A class inherited has been created to be container of parameters to the all widget's app. This class is parent of 'MyApp' in the main.dart file, that's mean any widget can reach the parameters like user, local and holding in the whole app.
Parameters and constructors has been deleted in widgets, because the inherited class is used to get them in thier context.
Dependencies updated of flutter_webview to use the clear cache method in the controller.
Updates in the user class, has been add the holding id and name in the web services.