Project moved to bloc pattern almost complete.
The item's page have been done and tested. But We have to fix a issue when you try to change the table, from "local" to "all locals", for a milliseconds the data loaded in the table are wrong, then refreshed by the stream builder and they're fine. Maybe bugs. Too, have been implemented a Button Navigator in the item's detail to split the content.